pyproject.toml
[project]
name = "pitot"
version = "1.0.0"
description = ""
authors = [
    {name = "Tianshu Huang",email = "thetianshuhuang@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "jax[cuda12]",
    "numpy",
    "dm-haiku",
    "optax",
    "pandas",
    "tqdm",
    "matplotlib",
    "tqdm-stubs",
    "jaxtyping",
    "beartype",
    "mypy",
    "scikit-learn"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
exclude = ["^plot/(.*).py"]

[[tool.mypy.overrides]]
module = ["optax.*"]
ignore_missing_imports = true
